mi: Remove questionable optimization from the rootless path
This is effectively a revert of 7b506fdc840aebed6b5acb91437a2cb620b5bddc
except the coding style reindent broke that.  The code makes no sense in
any case.  drawable can never be null since it's the first member of
WindowRec, and we're never called with a null window.  Neither can it be
an UNDRAWABLE_WINDOW since those are InputOnly windows; the rootless
code does not set the root window to either UNDRAWABLE or InputOnly, so.

Reviewed-by: Jasper St. Pierre <jstpierre@mecheye.net>
Signed-off-by: Adam Jackson <ajax@redhat.com>